refactor: extract grouping routes
