+++ src/routes/api.ts
@@ -4,286 +4,6 @@ import * as logger from '../logger';
 import { isProgramAdmin, isProgramMember } from "../utils/auth";
 const router = express.Router();
 
-
-router.post(
-  '/programs/:programId/grouping-types',
-  async (req: express.Request, res: express.Response) => {
-    const { programId } = req.params as { programId?: string };
-    const caller = (req as any).user as { userId: number; email: string };
-    if (!programId) {
-      res.status(400).json({ error: 'programId required' });
-      return;
-    }
-    const isAdmin = await isProgramAdmin(caller.userId, programId);
-    if (!isAdmin) {
-      res.status(403).json({ error: 'Forbidden' });
-      return;
-    }
-    const {
-      defaultName,
-      customName,
-      pluralName,
-      levelOrder,
-      isRequired,
-    } = req.body as {
-      defaultName?: string;
-      customName?: string;
-      pluralName?: string;
-      levelOrder?: number;
-      isRequired?: boolean;
-    };
-    if (!defaultName || levelOrder === undefined) {
-      res.status(400).json({ error: 'defaultName and levelOrder required' });
-      return;
-    }
-    const gt = await prisma.groupingType.create({
-      data: {
-        programId,
-        defaultName,
-        customName,
-        pluralName,
-        levelOrder,
-        isRequired: Boolean(isRequired),
-        status: 'active',
-      },
-    });
-    logger.info(programId, `GroupingType ${gt.id} created`);
-    res.status(201).json(gt);
-  },
-);
-
-router.get(
-  '/programs/:programId/grouping-types',
-  async (req: express.Request, res: express.Response) => {
-    const { programId } = req.params as { programId?: string };
-    const caller = (req as any).user as { userId: number };
-    if (!programId) {
-      res.status(400).json({ error: 'programId required' });
-      return;
-    }
-    const isMember = await isProgramMember(caller.userId, programId);
-    if (!isMember) {
-      res.status(403).json({ error: 'Forbidden' });
-      return;
-    }
-    const types = await prisma.groupingType.findMany({
-      where: { programId },
-      orderBy: { levelOrder: 'asc' },
-    });
-    res.json(types);
-  },
-);
-
-router.put('/grouping-types/:id', async (req: express.Request, res: express.Response) => {
-  const { id } = req.params as { id?: string };
-  const caller = (req as any).user as { userId: number };
-  const gt = await prisma.groupingType.findUnique({ where: { id: Number(id) } });
-  if (!gt) {
-    res.status(404).json({ error: 'Not found' });
-    return;
-  }
-  const isAdmin = await isProgramAdmin(caller.userId, gt.programId);
-  if (!isAdmin) {
-    res.status(403).json({ error: 'Forbidden' });
-    return;
-  }
-  const { customName, pluralName, levelOrder, isRequired, status } = req.body as {
-    customName?: string;
-    pluralName?: string;
-    levelOrder?: number;
-    isRequired?: boolean;
-    status?: string;
-  };
-  const updated = await prisma.groupingType.update({
-    where: { id: Number(id) },
-    data: {
-      customName,
-      pluralName,
-      levelOrder,
-      isRequired,
-      status,
-    },
-  });
-  logger.info(gt.programId, `GroupingType ${gt.id} updated`);
-  res.json(updated);
-});
-
-router.delete('/grouping-types/:id', async (req: express.Request, res: express.Response) => {
-  const { id } = req.params as { id?: string };
-  const caller = (req as any).user as { userId: number };
-  const gt = await prisma.groupingType.findUnique({ where: { id: Number(id) } });
-  if (!gt) {
-    res.status(404).json({ error: 'Not found' });
-    return;
-  }
-  const isAdmin = await isProgramAdmin(caller.userId, gt.programId);
-  if (!isAdmin) {
-    res.status(403).json({ error: 'Forbidden' });
-    return;
-  }
-  const updated = await prisma.groupingType.update({
-    where: { id: Number(id) },
-    data: { status: 'retired' },
-  });
-  logger.info(gt.programId, `GroupingType ${gt.id} retired`);
-  res.json(updated);
-});
-
-router.post('/programs/:programId/groupings', async (req: express.Request, res: express.Response) => {
-  const { programId } = req.params as { programId?: string };
-  const caller = (req as any).user as { userId: number; email: string };
-  if (!programId) {
-    res.status(400).json({ error: 'programId required' });
-    return;
-  }
-  const isAdmin = await isProgramAdmin(caller.userId, programId);
-  if (!isAdmin) {
-    res.status(403).json({ error: 'Forbidden' });
-    return;
-  }
-  const { groupingTypeId, parentGroupingId, name, displayOrder, notes } = req.body as {
-    groupingTypeId?: number;
-    parentGroupingId?: number;
-    name?: string;
-    displayOrder?: number;
-    notes?: string;
-  };
-  if (!groupingTypeId || !name) {
-    res.status(400).json({ error: 'groupingTypeId and name required' });
-    return;
-  }
-  const grouping = await prisma.grouping.create({
-    data: {
-      programId,
-      groupingTypeId,
-      parentGroupingId,
-      name,
-      displayOrder,
-      notes,
-      status: 'active',
-    },
-  });
-  logger.info(programId, `Grouping ${grouping.id} created`);
-  res.status(201).json(grouping);
-});
-
-router.get('/programs/:programId/groupings', async (req: express.Request, res: express.Response) => {
-  const { programId } = req.params as { programId?: string };
-  const caller = (req as any).user as { userId: number };
-  if (!programId) {
-    res.status(400).json({ error: 'programId required' });
-    return;
-  }
-  const isMember = await isProgramMember(caller.userId, programId);
-  if (!isMember) {
-    res.status(403).json({ error: 'Forbidden' });
-    return;
-  }
-  const groupings = await prisma.grouping.findMany({
-    where: { programId },
-    orderBy: { displayOrder: 'asc' },
-  });
-  res.json(groupings);
-});
-
-router.put('/groupings/:id', async (req: express.Request, res: express.Response) => {
-  const { id } = req.params as { id?: string };
-  const caller = (req as any).user as { userId: number };
-  const grouping = await prisma.grouping.findUnique({ where: { id: Number(id) } });
-  if (!grouping) {
-    res.status(404).json({ error: 'Not found' });
-    return;
-  }
-  const isAdmin = await isProgramAdmin(caller.userId, grouping.programId);
-  if (!isAdmin) {
-    res.status(403).json({ error: 'Forbidden' });
-    return;
-  }
-  const { name, displayOrder, notes, parentGroupingId, status } = req.body as {
-    name?: string;
-    displayOrder?: number;
-    notes?: string;
-    parentGroupingId?: number;
-    status?: string;
-  };
-  const updated = await prisma.grouping.update({
-    where: { id: Number(id) },
-    data: { name, displayOrder, notes, parentGroupingId, status },
-  });
-  logger.info(grouping.programId, `Grouping ${grouping.id} updated`);
-  res.json(updated);
-});
-
-router.delete('/groupings/:id', async (req: express.Request, res: express.Response) => {
-  const { id } = req.params as { id?: string };
-  const caller = (req as any).user as { userId: number };
-  const grouping = await prisma.grouping.findUnique({ where: { id: Number(id) } });
-  if (!grouping) {
-    res.status(404).json({ error: 'Not found' });
-    return;
-  }
-  const isAdmin = await isProgramAdmin(caller.userId, grouping.programId);
-  if (!isAdmin) {
-    res.status(403).json({ error: 'Forbidden' });
-    return;
-  }
-  const updated = await prisma.grouping.update({
-    where: { id: Number(id) },
-    data: { status: 'retired' },
-  });
-  logger.info(grouping.programId, `Grouping ${grouping.id} retired`);
-  res.json(updated);
-});
-
-router.post('/program-years/:id/groupings/activate', async (req: express.Request, res: express.Response) => {
-  const { id } = req.params as { id?: string };
-  const caller = (req as any).user as { userId: number };
-  const py = await prisma.programYear.findUnique({ where: { id: Number(id) } });
-  if (!py) {
-    res.status(404).json({ error: 'Not found' });
-    return;
-  }
-  const isAdmin = await isProgramAdmin(caller.userId, py.programId);
-  if (!isAdmin) {
-    res.status(403).json({ error: 'Forbidden' });
-    return;
-  }
-  const { groupingIds } = req.body as { groupingIds?: number[] };
-  if (!Array.isArray(groupingIds) || groupingIds.length === 0) {
-    res.status(400).json({ error: 'groupingIds required' });
-    return;
-  }
-  const records = await Promise.all(
-    groupingIds.map((gid) =>
-      prisma.programYearGrouping.create({
-        data: { programYearId: py.id, groupingId: gid, status: 'active' },
-      })
-    )
-  );
-  logger.info(py.programId, `Activated ${records.length} groupings for PY ${py.year}`);
-  res.status(201).json(records);
-});
-
-router.get('/program-years/:id/groupings', async (req: express.Request, res: express.Response) => {
-  const { id } = req.params as { id?: string };
-  const caller = (req as any).user as { userId: number };
-  const py = await prisma.programYear.findUnique({ where: { id: Number(id) } });
-  if (!py) {
-    res.status(404).json({ error: 'Not found' });
-    return;
-  }
-  const isMember = await isProgramMember(caller.userId, py.programId);
-  if (!isMember) {
-    res.status(403).json({ error: 'Forbidden' });
-    return;
-  }
-  const records = await prisma.programYearGrouping.findMany({
-    where: { programYearId: py.id, status: 'active' },
-    include: { grouping: true },
-  });
-  res.json(records);
-});
-
 router.post('/programs/:programId/parties', async (req: express.Request, res: express.Response) => {
   const { programId } = req.params as { programId?: string };
   const caller = (req as any).user as { userId: number; email: string };
